check the bug

+++ service/hyjj_service/src/main/java/com/hyjj/hyjjservice/controller/fill/dto/FormulaVerificationDto.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public class FormulaVerificationDto {
     private Long reportId;
-    private List<Long> data;
+    private List<Double> data;
 
     public Long getReportId() {
         return reportId;
@@ -14,11 +14,11 @@ public class FormulaVerificationDto {
         this.reportId = reportId;
     }
 
-    public List<Long> getData() {
+    public List<Double> getData() {
         return data;
     }
 
-    public void setData(List<Long> data) {
+    public void setData(List<Double> data) {
         this.data = data;
     }
 }
