+++ service/hyjj_service/src/main/java/com/hyjj/hyjjservice/service/statistic/impl/template/TemplateId25.java
@@ -31,7 +31,7 @@ public class TemplateId25 extends AbstractTargetTemplate {
     }
 
     @Override
-    public Boolean formulaVerification(List<Long> data) {
+    public Boolean formulaVerification(List<Double> data) {
         return data.get(0) >= data.get(1);
     }
 }
