+++ service/hyjj_service/src/main/java/com/hyjj/hyjjservice/service/statistic/impl/template/TemplateId30.java
@@ -33,7 +33,7 @@ public class TemplateId30 extends AbstractTargetTemplate {
     }
 
     @Override
-    public Boolean formulaVerification(List<Long> data) {
+    public Boolean formulaVerification(List<Double> data) {
         return data.get(0) >= data.get(1) &&
                 data.get(2) >= data.get(3) &&
                 data.get(4) >= data.get(5);
